Fix all compile-blocking errors across compute, governance, orchestration, services, httpapi, and app packages
- orchestration/nodeagent.go: fix DeployWorkload to use Workload.Spec.* fields; remove duplicate scheduleWorkload method and orphaned ReplicaState/NodeInfo types
- orchestration/prometheus.go: fix WorkloadState field access (use .Workload.*), PendingQueue len, node field names (DID/TotalCPU/etc); use discovery.ListNodes()
- orchestration/discovery.go: add ListNodes() method for prometheus exporter
- orchestration/scheduler_api.go: add FedScheduler API shim methods (ListWorkloads, GetWorkload, ScaleWorkload, DeleteWorkload, UpdateWorkload, RestartWorkload, GetWorkloadLogs, GetWorkloadMetrics, GetWorkloadEvents)
- services/prometheus.go: fix ServiceCatalog→Catalog, ContainerStats field refs, ServiceMetrics field names; rewrite to use actual types
- services/catalog.go: add ListAllInstances() for prometheus exporter
- services/backup.go: replace undefined fileExists() with os.Stat check
- services/mongodb,mysql,objectstore,postgres,queue,redis: fix plan.CPU→plan.CPUCores
- httpapi/server.go: fix orchestration.Constraints→PlacementConstraints; fix WorkloadState field access
- httpapi/revenue.go: fix int64→float64 casts for Prometheus revenue maps; fix pending balance comparison
- httpapi/services.go: fix ServiceTypeQueue→ServiceTypeMessageQueue; fix ProvisionRequest field names (InstanceID→Name, PlanID→Plan); add missing time/strconv imports; fix Credentials struct access
- httpapi/workloads.go: remove unused orchestration import; fix unused workloadID variable
- app/app.go: fix payment constructor calls (add store arg); fix service provisioner calls (add docker endpoint arg)
- security-fixes/*.go: add //go:build ignore to prevent package conflict
- cmd/soholink-wizard/main.go: add //go:build gui tag to exclude from non-CGO builds

+++ internal/httpapi/revenue.go
@@ -105,15 +105,15 @@ func (s *Server) handleGetRevenueStats(w http.ResponseWriter, r *http.Request) {
 	json.NewEncoder(w).Encode(map[string]interface{}{
 		"total": total,
 		"periods": map[string]float64{
-			"last_24h": dayRevenue,
-			"last_7d":  weekRevenue,
-			"last_30d": monthRevenue,
+			"last_24h": float64(dayRevenue),
+			"last_7d":  float64(weekRevenue),
+			"last_30d": float64(monthRevenue),
 		},
 		"by_resource_type": map[string]float64{
-			"cpu":     cpuRevenue,
-			"memory":  memoryRevenue,
-			"storage": storageRevenue,
-			"gpu":     gpuRevenue,
+			"cpu":     float64(cpuRevenue),
+			"memory":  float64(memoryRevenue),
+			"storage": float64(storageRevenue),
+			"gpu":     float64(gpuRevenue),
 		},
 	})
 }
@@ -175,7 +175,7 @@ func (s *Server) handleRequestPayout(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if req.Amount > pending {
+	if req.Amount > float64(pending) {
 		http.Error(w, "Insufficient pending balance", http.StatusBadRequest)
 		return
 	}
